requirements.txt
Flask>=2.0
Flask-SocketIO
gunicorn
eventlet
psycopg2-binary
requests
authlib
Werkzeug
